add retract block to go.mod files
This marks the incorrectly tagged version as retracted, which means go
tooling will no longer consider it the latest version (and won't attempt
to automatically upgrade to it).

In addition, once this is merged, we will need to tag v1.2.4:
- this creates a new largest semver version, which go will consider as
  latest
- go tooling will download this v1.2.4 version, read the retract
  statements, then correctly interpret the true latest version to be
  v0.X.Y

This is needed because once a version is added to the sum.golang.org
database, it is never removed (prioritizing build stability). See
https://go.dev/ref/mod#checksum-database, which this change is based
off, for more details.